pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "rag-saas"
version = "0.1.0"
description = "Enterprise-grade multi-tenant RAG SaaS platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.0.0",
    "arq>=0.26.0",
    "litellm>=1.50.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=41.0.0",
    "svix>=1.0.0",
    # Document ingestion pipeline
    "minio>=7.2.0",
    "langchain-text-splitters>=0.3.0",
    "tiktoken>=0.7.0",
    # Generation pipeline
    "sse-starlette>=2.0.0",
    # Billing
    "stripe>=10.0.0",
    # Connectors
    "google-auth>=2.28.0",
    "google-api-python-client>=2.120.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "testcontainers[postgres]>=4.0.0",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
